requirements.txt
autopep8==1.5.7
bleach==4.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
docutils==0.17.1
idna==3.2
importlib-metadata==4.8.0
keyring==23.1.0
packaging==21.0
pkginfo==1.7.1
pycodestyle==2.7.0
Pygments==2.10.0
pyparsing==2.4.7
python-dotenv==0.19.0
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
toml==0.10.2
tqdm==4.62.2
twine==3.4.2
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
